+++ examples/plots/archive/colors.py
