+++ examples/plots/archive/fam_acc.png
