+++ examples/plots/archive/isolates.pdf
