+++ examples/plots/archive/isolates.png
