+++ examples/plots/archive/neighbornet.stree6
