+++ examples/plots/archive/scatter.png
