+++ examples/plots/archive/scope.png
