+++ examples/plots/archive/violin_complex.png
