+++ examples/plots/isolate-plots.py
