+++ examples/plots/violin-plots.py
