+++ examples/plots/violins.py
