--- examples/requirements.txt
@@ -1,104 +1,0 @@
--e git+ssh://git@github.com/lexibank/ala@b69197ba3d934c8b50b9fbca735c6206e9a1056c#egg=ala
-appdirs==1.4.4
-asttokens==3.0.0
-attrs==25.1.0
-babel==2.17.0
-bibtexparser==2.0.0b8
-cdstarcat==1.5.0
-certifi==2025.1.31
-charset-normalizer==3.4.1
-cldfcatalog==1.5.1
-clldutils==3.24.1
-colorama==0.4.6
-colorlog==6.9.0
-commonnexus==1.9.2
-contourpy==1.3.1
-csvw==3.5.1
-cycler==0.12.1
-decorator==5.1.1
-docopt==0.6.2
-executing==2.2.0
-filelock==3.17.0
-fonttools==4.56.0
-fsspec==2025.2.0
-gitdb==4.0.12
-GitPython==3.1.44
-idna==3.10
-ipython==8.32.0
-isodate==0.7.2
-jedi==0.19.2
-Jinja2==3.1.5
-jmespath==1.0.1
-joblib==1.4.2
-jsonschema==4.23.0
-jsonschema-specifications==2024.10.1
-kiwisolver==1.4.8
-language-tags==1.2.0
-latexcodec==3.0.0
-lightning-utilities==0.12.0
-lingpy==2.6.13
-lxml==5.3.1
-Markdown==3.7
-MarkupSafe==3.0.2
-matplotlib==3.10.0
-matplotlib-inline==0.1.7
-mpmath==1.3.0
-nameparser==1.1.3
-networkx==3.4.2
-newick==1.9.0
-numpy==2.2.2
-nvidia-cublas-cu12==12.4.5.8
-nvidia-cuda-cupti-cu12==12.4.127
-nvidia-cuda-nvrtc-cu12==12.4.127
-nvidia-cuda-runtime-cu12==12.4.127
-nvidia-cudnn-cu12==9.1.0.70
-nvidia-cufft-cu12==11.2.1.3
-nvidia-curand-cu12==10.3.5.147
-nvidia-cusolver-cu12==11.6.1.9
-nvidia-cusparse-cu12==12.3.1.170
-nvidia-cusparselt-cu12==0.6.2
-nvidia-nccl-cu12==2.21.5
-nvidia-nvjitlink-cu12==12.4.127
-nvidia-nvtx-cu12==12.4.127
-packaging==24.2
-parso==0.8.4
-pexpect==4.9.0
-pillow==11.1.0
-prompt_toolkit==3.0.50
-ptyprocess==0.7.0
-pure_eval==0.2.3
-pybtex==0.24.0
-pycdstar==1.1.0
-pycldf==1.40.4
-pyconcepticon==3.1.0
-Pygments==2.19.1
-pylatexenc==2.10
-pyparsing==3.2.1
-python-dateutil==2.9.0.post0
-python-frontmatter==1.1.0
-PyYAML==6.0.2
-rdflib==7.1.3
-referencing==0.36.2
-requests==2.32.3
-rfc3986==1.5.0
-rpds-py==0.22.3
-scikit-learn==1.6.1
-scipy==1.15.1
-setuptools==75.8.0
-six==1.17.0
-smmap==5.0.2
-stack-data==0.6.3
-sympy==1.13.1
-tabulate==0.9.0
-termcolor==2.5.0
-threadpoolctl==3.5.0
-torch==2.6.0
-torchmetrics==1.6.1
-tqdm==4.67.1
-traitlets==5.14.3
-triton==3.2.0
-typing_extensions==4.12.2
-Unidecode==1.3.8
-uritemplate==4.1.1
-urllib3==2.3.0
-wcwidth==0.2.13
