--- examples/runs-bl.sh
@@ -1,8 +1,0 @@
-python baseline.py 100 asjp
-python baseline.py 100 lb Swadesh-1952-200
-python baseline.py 100 lb Swadesh-1955-100
-python baseline.py 100 lb Tadmor-2009-100
-python baseline.py 100 lb Starostin-1991-110
-
-
-
