--- examples/runs-nn.sh
@@ -1,7 +1,0 @@
-python ala-run.py --database=lexibank --runs=100
-python ala-run.py --database=grambank --runs=100
-python ala-run.py --datasbase=combined --runs=100
-python ala-run.py --database=asjp --runs=100
-
-
-
