+++ examples/runs_bl.sh
@@ -1,0 +1,5 @@
+python ala_baseline.py 100 asjp
+python ala_baseline.py 100 lb Swadesh-1952-200
+python ala_baseline.py 100 lb Swadesh-1955-100
+python ala_baseline.py 100 lb Tadmor-2009-100
+python ala_baseline.py 100 lb Starostin-1991-110
