+++ examples/runs_comparison_nn.sh
@@ -1,0 +1,4 @@
+python ala_nn_asjp.py --runs=100
+python ala_nn_combined.py --runs=100
+python ala_nn_grambank.py --runs=100
+python ala_nn_lexibank.py --runs=100
